Revert "[kimi-k2.5]vit部分做w8a8_mxfp8量化时，量化权重被强转bug"
This reverts commit cc208f3116b395aab96ff2032a3c77d164baa5d3.

The dtype override issue is now handled by the vllm-ascend patch
(MoonViT3dPretrainedModel.to monkey-patch) which strips the dtype
argument at runtime, making the in-source workaround unnecessary.

+++ vllm/model_executor/models/kimi_k25.py
@@ -375,10 +375,9 @@ class KimiK25ForConditionalGeneration(
                 quant_config=self._maybe_ignore_quant_config(quant_config),
                 prefix=maybe_prefix(prefix, "vision_tower"),
             )
-            # self.vision_tower = self.vision_tower.to(
-            #     device=self.device, dtype=model_config.dtype
-            # )
-            self.vision_tower = self.vision_tower.to(device=self.device)
+            self.vision_tower = self.vision_tower.to(
+                device=self.device, dtype=model_config.dtype
+            )
 
             self.mm_projector = KimiK25MultiModalProjector(
                 config=config.vision_config,
